+++ include/ftxui/util/ref.hpp
@@ -48,6 +48,7 @@ class StringRef {
   StringRef(const wchar_t* ref) : StringRef(to_string(std::wstring(ref))) {}
   StringRef(const char* ref) : StringRef(std::string(ref)) {}
   std::string& operator*() { return address_ ? *address_ : owned_; }
+  std::string& operator()() { return address_ ? *address_ : owned_; }
   std::string* operator->() { return address_ ? address_ : &owned_; }
 
  private: