+++ src/ftxui/component/receiver_test.cpp
@@ -1,8 +1,9 @@
-#include <gtest/gtest.h>
+#include <string>   // for string
 #include <thread>   // for thread
 #include <utility>  // for move
 
 #include "ftxui/component/receiver.hpp"
+#include "gtest/gtest.h"  // for AssertionResult, Message, Test, TestPartResult, EXPECT_EQ, EXPECT_TRUE, EXPECT_FALSE, TEST
 
 // NOLINTBEGIN
 namespace ftxui {