+++ src/ftxui/dom/blink_test.cpp
@@ -1,9 +1,9 @@
-#include <gtest/gtest.h>
-#include <string>  // for allocator
+#include <string>  // for allocator, string
 
 #include "ftxui/dom/elements.hpp"   // for operator|, text, blink, Element
 #include "ftxui/dom/node.hpp"       // for Render
 #include "ftxui/screen/screen.hpp"  // for Screen, Pixel
+#include "gtest/gtest.h"  // for Test, AssertionResult, EXPECT_TRUE, Message, TEST, TestPartResult
 
 // NOLINTBEGIN
 namespace ftxui {